Create hexToRGB helper
We were using a mix of #FFFFFF and 255,255,255 notation in the colors enum. The latter was needed by a gradient on some cards that uses RGBA to control fade. It was not a valid format to use on its own though. to avoid duplication everything now follows #FFFFFF format and there is a helper to convert when needed.