Begin creating all-in-one 15 game page
The hope was that this could reduce code duplication but it is probably foolish. It may be worth while to lift certain pieces of state like the WLD record but there are many slight differences between playing vs a bot or a human - this 'consolidation would result in a very large file with many of the same functions we currently have.